ChatRoom 권한 검증 추가가

+++ 02-backend/00-monolitic/src/main/kotlin/site/rahoon/message/__monolitic/chatroom/application/ChatRoomApplicationService.kt
@@ -3,7 +3,9 @@ package site.rahoon.message.__monolitic.chatroom.application
 import org.springframework.stereotype.Service
 import site.rahoon.message.__monolitic.chatroom.domain.ChatRoomCommand
 import site.rahoon.message.__monolitic.chatroom.domain.ChatRoomDomainService
+import site.rahoon.message.__monolitic.chatroom.domain.ChatRoomError
 import site.rahoon.message.__monolitic.chatroom.domain.ChatRoomInfo
+import site.rahoon.message.__monolitic.common.domain.DomainException
 
 /**
  * ChatRoom Application Service
@@ -17,8 +19,8 @@ class ChatRoomApplicationService(
     /**
      * 채팅방 생성
      */
-    fun create(criteria: ChatRoomCriteria.Create, createdByUserId: String): ChatRoomInfo.Detail {
-        val command = criteria.toCommand(createdByUserId)
+    fun create(criteria: ChatRoomCriteria.Create): ChatRoomInfo.Detail {
+        val command = criteria.toCommand()
         // TODO ChatRoomUser에 본인 추가 필요.
         return chatRoomDomainService.create(command)
     }
@@ -26,16 +28,42 @@ class ChatRoomApplicationService(
     /**
      * 채팅방 수정
      */
-    fun update(criteria: ChatRoomCriteria.Update, chatRoomId: String): ChatRoomInfo.Detail {
-        val command = criteria.toCommand(chatRoomId)
+    fun update(criteria: ChatRoomCriteria.Update): ChatRoomInfo.Detail {
+        // 권한 검증: 생성자만 수정 가능
+        val chatRoomInfo = chatRoomDomainService.getById(criteria.chatRoomId)
+        if (chatRoomInfo.createdByUserId != criteria.userId) {
+            throw DomainException(
+                error = ChatRoomError.UNAUTHORIZED_ACCESS,
+                details = mapOf(
+                    "chatRoomId" to criteria.chatRoomId,
+                    "userId" to criteria.userId,
+                    "reason" to "채팅방 생성자만 수정할 수 있습니다"
+                )
+            )
+        }
+
+        val command = criteria.toCommand()
         return chatRoomDomainService.update(command)
     }
 
     /**
      * 채팅방 삭제
      */
-    fun delete(chatRoomId: String): ChatRoomInfo.Detail {
-        val command = ChatRoomCommand.Delete(id = chatRoomId)
+    fun delete(criteria: ChatRoomCriteria.Delete): ChatRoomInfo.Detail {
+        // 권한 검증: 생성자만 삭제 가능
+        val chatRoomInfo = chatRoomDomainService.getById(criteria.chatRoomId)
+        if (chatRoomInfo.createdByUserId != criteria.userId) {
+            throw DomainException(
+                error = ChatRoomError.UNAUTHORIZED_ACCESS,
+                details = mapOf(
+                    "chatRoomId" to criteria.chatRoomId,
+                    "userId" to criteria.userId,
+                    "reason" to "채팅방 생성자만 삭제할 수 있습니다"
+                )
+            )
+        }
+
+        val command = criteria.toCommand()
         return chatRoomDomainService.delete(command)
     }
